HDDS-825. Code cleanup based on messages from ErrorProne. Contributed by Anu Engineer.

+++ hadoop-ozone/tools/src/main/java/org/apache/hadoop/ozone/freon/RandomKeyGenerator.java
@@ -286,7 +286,7 @@ public final class RandomKeyGenerator implements Callable<Void> {
     long maxValue;
 
     currentValue = () -> numberOfKeysAdded.get();
-    maxValue = numOfVolumes *
+    maxValue = (long) numOfVolumes *
         numOfBuckets *
         numOfKeys;
